requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Model Interpretation
shap>=0.40.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.4.0
notebook>=6.4.0

# Data Processing
openpyxl>=3.0.9  # For Excel files
pyyaml>=6.0  # For config files

# Utilities
tqdm>=4.62.0  # Progress bars
python-dateutil>=2.8.0

# Optional: Deep Learning (uncomment if needed)
# tensorflow>=2.8.0
# torch>=1.10.0
